Fix ButtonModel 2

+++ source/ButtonModel.cpp
@@ -12,7 +12,8 @@ void ButtonModel::dispose() {}
 
 int ButtonModel::getSection() {
 	float mod = fmod(getAngle(), (float)globals::SEG_DEG);
-	int section = (int)(mod < (globals::SEG_DEG / 2.0f) ? ceilf(getAngle() / globals::SEG_DEG)
-														: floorf(getAngle() / globals::SEG_DEG));
+	int section =
+		(int)(mod < ((float)globals::SEG_DEG / 2) ? ceilf(getAngle() / globals::SEG_DEG)
+												  : floorf(getAngle() / globals::SEG_DEG));
 	return section;
 }
